Unlink sign up and handler to sign in handler
